Задачи №6 Исправил, выполнил в условии преобразование к целому.

+++ src/main/Task06/Calc.java
@@ -13,7 +13,7 @@ public class Calc {
         z=in.nextInt();
         avg= (float)((x+y+z)/3.0);
         System.out.println("Среднее арифметическое x, y, z: " +avg) ;
-        if (avg/2>3) {
+        if ((int)avg/2>3) {
             System.out.println("Программа выполнена корректно");
         }
     }
